Refactor record generation by implementing ModelFactory and updating RecordGenerator for improved model handling
